feat: add javadoc for module Dispatcher
