--- dispatcher/src/main/java/com/github/jon7even/cache/UserAuthCache.java
@@ -1,9 +1,0 @@
-package com.github.jon7even.cache;
-
-public interface UserAuthCache {
-    void setAttemptAuthForUserCache(Long userId, Integer attemptAuth);
-
-    int getAttemptsAuthForUserCache(Long userId);
-
-    void deleteUserFromAuthCache(Long userId);
-}
