+++ dispatcher/src/main/java/com/github/jon7even/configuration/BotConfig.java
@@ -5,6 +5,12 @@ import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * Утилитарный класс загружающий токен и имя бота из переменных окружения
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Getter
 @Setter
 @Configuration
