+++ dispatcher/src/main/java/com/github/jon7even/configuration/RabbitConfiguration.java
@@ -8,6 +8,12 @@ import org.springframework.context.annotation.Configuration;
 
 import static com.github.jon7even.configuration.RabbitQueue.*;
 
+/**
+ * Конфигурация для работы с брокером очередей RabbitMq
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Configuration
 public class RabbitConfiguration {
     @Bean
