+++ dispatcher/src/main/java/com/github/jon7even/dto/UserShortDto.java
@@ -6,7 +6,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 
-
+/**
+ * Класс DTO для краткого представления пользователя
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Data
 @Builder
 @AllArgsConstructor
