+++ dispatcher/src/main/java/com/github/jon7even/service/in/AnswerConsumerServiceImpl.java
@@ -1,4 +1,4 @@
-package com.github.jon7even.service.out;
+package com.github.jon7even.service.in;
 
 import com.github.jon7even.controller.out.SenderBotClient;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +9,12 @@ import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 
 import static com.github.jon7even.configuration.RabbitQueue.ANSWER_MESSAGE;
 
+/**
+ * Реализация сервиса слушателя RabbitMq для отправки сообщений в Telegram bot
+ *
+ * @author Jon7even
+ * @version 1.0
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
