--- dispatcher/src/main/java/com/github/jon7even/service/in/AuthorizationService.java
@@ -1,9 +1,0 @@
-package com.github.jon7even.service.in;
-
-import com.github.jon7even.dto.UserShortDto;
-
-public interface AuthorizationService {
-    boolean processAuthorization(UserShortDto userShortDto);
-
-    boolean processAuthorizationForCallBack(Long userId);
-}
